requirements.txt
tensorflow
numpy
pandas
matplotlib
seaborn
scikit-learn
streamlit
pillow
